requirements.txt
aenum==3.1.15
aioredis==1.3.1
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
cachetools==5.5.2
cashfree-pg==4.3.10
cashfree_sdk==0.1.3
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.2
daphne==2.5.0
distlib==0.3.9
Django==5.1.7
django-allauth==65.5.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-filter==23.2
django-redis==5.4.0
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
hiredis==3.1.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
kombu==5.5.3
msgpack==1.1.0
oauthlib==3.2.2
pem==23.1.0
pillow==11.1.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==1.10.21
PyJWT==2.9.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
razorpay==1.4.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
sentry-sdk==1.32.0
service-identity==24.2.0
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.0.7
vine==5.1.0
virtualenv==20.29.3
wcwidth==0.2.13
zope.interface==7.2
